Factored out all address-bar-management logic into a separate module.
The new file is called current-page-manager.js.
